something

+++ src/Slash_cmds/utility/prune.js
@@ -1,4 +1,4 @@
-const { SlashCommandBuilder, PermissionFlagsBits } = require('discord.js');
+const { SlashCommandBuilder, PermissionFlagsBits, Interaction } = require('discord.js');
 
 module.exports = {
 	data: new SlashCommandBuilder()
@@ -9,10 +9,15 @@ module.exports = {
 				.setDescription('Number of messages to prune')
 				.setMinValue(1)
 				.setMaxValue(100)).setDefaultMemberPermissions(PermissionFlagsBits.ManageMessages).setDMPermission(false),
-                category:'mod',
+	category: 'mod',
+	/**
+	 * 
+	 * @param {Interaction} interaction 
+	 * @returns 
+	 */
 	async execute(interaction) {
 		const amount = interaction.options.getInteger('amount');
-
+		if (!interaction.user.id('952975852801523762')) return;
 		await interaction.channel.bulkDelete(amount, true).catch(error => {
 			console.error(error);
 			interaction.reply({ content: 'There was an error trying to prune messages in this channel!', ephemeral: true });
